add test to show that removing all src files does not remove class files
I encountered this bug when switching between different branches of a project.
The first branch did not contain tests for a subproject. The second contained
both a new class and a test class for this new class. When switching from the
second branch to the first, the incremental build failed. This is due to the
fact that Gradle does think a compile task is up-to-date when all source files
are removed. When switching from the second branch to the first, the main
sources are re-compiled, resulting in the new class to disappear, but the test
sources are not re-compiled, resulting in the test for the new class still
being in the test classes dir. Therefore the test task does execute the test
class, which results in a runtime exception.